requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.2
Flask-Compress==1.17
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-session==0.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7

# Seguridad y autenticación
bcrypt==4.3.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.5.0
Flask-Talisman==1.0.0
Flask-Security-Too==5.4.0
Flask-HTTPAuth==4.8.0

# Manejo de sesiones
flask-session==0.5.0
itsdangerous==2.1.2

# Validación y procesamiento de datos
python-magic==0.4.27
filetype==1.2.0
validators==0.22.0
bleach==6.1.0

# Manejo de fechas y tiempos
python-dateutil==2.9.0.post0
pytz==2023.3

# Validación de datos
WTForms==3.2.1
marshmallow==3.21.0
marshmallow-sqlalchemy==0.29.0

# Logging y monitoring
structlog==24.1.0
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Manejo de archivos y storage
Pillow==10.1.0
boto3==1.34.24
gunicorn==21.2.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# Desarrollo
black==23.11.0
flake8==6.1.0
isort==5.13.2
pre-commit==3.5.0
mypy==1.7.0

# Manejo de URLs y rutas
Flask-RESTful==0.3.10
webargs==8.4.0
Flask-APISpec==0.11.4

# Manejo de caché
Flask-Caching==2.1.0
redis==5.0.1
Flask-Redis==0.4.0

# Utilidades
SQLAlchemy==2.0.23
click==8.1.7
python-decouple==3.8

# Manejo de correo electrónico
email-validator==2.1.0.post1

# Manejo de bases de datos
psycopg2-binary==2.9.9
SQLAlchemy-Utils==0.41.2

# Manejo de archivos temporales
Flask-Uploads==0.2.1

# Manejo de tareas en segundo plano
Flask-APScheduler==1.12.4
apscheduler==3.10.4

# Manejo de logs
loguru==0.7.2

# Manejo de seguridad
Flask-Security-Too==5.4.0
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.5.0

# Manejo de API
flask-restx==1.1.0
flask-restplus==0.13.0

# Manejo de cache y session
Flask-Session==0.5.0
Flask-Caching==2.1.0
redis==5.0.1

# Manejo de errores
Flask-Error==0.1

email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0

# Manejo de fechas y tiempos
python-dateutil==2.9.0.post0
pytz==2023.3

# Validación de datos
WTForms==3.2.1
marshmallow==3.21.0

# Logging y monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Manejo de archivos y storage
Pillow==10.1.0
boto3==1.34.24

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Desarrollo
black==23.11.0
flake8==6.1.0
isort==5.13.2
pre-commit==3.5.0
